\name{air_document}
\alias{air_document}
\title{air_document}
\description{
  Function air_document from the raix package.
}
\usage{
  air_document(code, func_name = NULL)
}
\arguments{
  \item{code}{Argument.}
  \item{func_name}{Argument.}
}
\examples{
\dontrun{
  air_document()
}
}
\keyword{internal}
