\name{air_generate}
\alias{air_generate}
\title{air_generate}
\description{
  Function air_generate from the raix package.
}
\usage{
  air_generate(description, context = NULL)
}
\arguments{
  \item{description}{Argument.}
  \item{context}{Argument.}
}
\examples{
\dontrun{
  air_generate()
}
}
\keyword{internal}
