\name{air_google}
\alias{air_google}
\title{air_google}
\description{
  Function air_google from the raix package.
}
\usage{
  air_google(query, open_browser = TRUE)
}
\arguments{
  \item{query}{Argument.}
  \item{open_browser}{Argument.}
}
\examples{
\dontrun{
  air_google()
}
}
\keyword{internal}
