\name{air_rstudio}
\alias{air_rstudio}
\title{air_rstudio}
\description{
  Function air_rstudio from the raix package.
}
\usage{
  air_rstudio(mode = c("pane", "source", "console"))
}
\arguments{
  \item{mode}{Display mode: "pane" (Viewer), "source" (editor), or "console".}
}
\examples{
\dontrun{
  air_rstudio()
}
}
\keyword{internal}
