\name{air_send}
\alias{air_send}
\title{air_send}
\description{
  Function air_send from the raix package.
}
\usage{
  air_send(prompt, context = NULL, stream = FALSE)
}
\arguments{
  \item{prompt}{Argument.}
  \item{context}{Argument.}
  \item{stream}{Argument.}
}
\examples{
\dontrun{
  air_send()
}
}
\keyword{internal}
