\name{raix_document}
\alias{raix_document}
\title{raix_document}
\description{
  Function raix_document from the raix package.
}
\usage{
  raix_document(code, func_name = NULL)
}
\arguments{
  \item{code}{Argument.}
  \item{func_name}{Argument.}
}
\examples{
\dontrun{
  raix_document()
}
}
\keyword{internal}
