\name{raix_generate}
\alias{raix_generate}
\title{raix_generate}
\description{
  Function raix_generate from the raix package.
}
\usage{
  raix_generate(description, context = NULL)
}
\arguments{
  \item{description}{Argument.}
  \item{context}{Argument.}
}
\examples{
\dontrun{
  raix_generate()
}
}
\keyword{internal}
