\name{raix_google}
\alias{raix_google}
\title{raix_google}
\description{
  Function raix_google from the raix package.
}
\usage{
  raix_google(query, open_browser = TRUE)
}
\arguments{
  \item{query}{Argument.}
  \item{open_browser}{Argument.}
}
\examples{
\dontrun{
  raix_google()
}
}
\keyword{internal}
