\name{raix_send}
\alias{raix_send}
\title{raix_send}
\description{
  Function raix_send from the raix package.
}
\usage{
  raix_send(prompt, context = NULL, stream = FALSE)
}
\arguments{
  \item{prompt}{Argument.}
  \item{context}{Argument.}
  \item{stream}{Argument.}
}
\examples{
\dontrun{
  raix_send()
}
}
\keyword{internal}
